% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eTaxtools.R, R/write.eTax.R
\docType{package}
\name{write.eTax}
\alias{write.eTax}
\title{eTaxtools}
\usage{
write.eTax(f)
}
\arguments{
\item{f}{f(.xlsx)　→ f(.csv)}
}
\description{
Put eTax csv from xlsx

xlsxからCSV（eTax型式）を作成する。
}
